core/requirements.txt
# celery>=5.0.5
redis>=3.5.3
psycopg2-binary==2.8.5
numpy
pandas